+++ NEWS.md
@@ -1,8 +1,20 @@
+# ALASCA 0.0.0.96
+
+## New features
+
+* `assessGroupDifferences()` now supports `lm()`
+* One can now use Cox regression to take limits of detection into account. To do this, specify `method = "KM"` (using `survival::coxph()`) for ASCA or `method = "KMM"` (using `coxme::coxme()`) for RM-ASCA. You should be careful with scaling; `scaleFun <- function(x){return(x)}` is an alternative. Next, you need to make a data frame and provide it to the function `lowerLimit = lowerLimit`. `lowerLimit` should have a column `value` with the lower cutoff values, and `variable` for the variable names (must correspond to the variables in `df`!)
+  * Note: This function is in alpha and has not been thoroughly tested!
+
+## Breaking changes
+
+* The choice to use `Rfast` is now moved to a separate parameter: `useRfast` which defaults to `TRUE`. The goal is to implement `Rfast` for linear models as well.
+
 # ALASCA 0.0.0.95
 
 ## New features
 
-* `assessGroupDifferences` applies `emmeans::emmeans()` to calculate pairwise group differences
+* `assessGroupDifferences()` applies `emmeans::emmeans()` to calculate pairwise group differences
 
 # ALASCA 0.0.0.94